starter-packs/speech-to-text-pack/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-to-text-pack"
version = "1.0.0"
description = "Transcribe audio to text using OpenAI Whisper (local or API)."
requires-python = ">=3.10"
license = "MIT"
dependencies = ["openai-whisper>=20231117", "httpx>=0.25"]

[tool.setuptools.packages.find]
where = ["src"]
